Apply code suggestions * renamed fuse2 package and maven execution * remove fuse3_loop_config * improve doc
