+++ jfuse-win-amd64/pom.xml
@@ -115,7 +115,6 @@
 										<includeStruct>fuse_statvfs</includeStruct>
 										<includeStruct>fuse_timespec</includeStruct>
 										<includeStruct>fuse3_conn_info</includeStruct>
-										<includeStruct>fuse3_loop_config</includeStruct>
 									</includeStructs>
 									<includeMacros>
 										<includeMacro>FUSE_FILL_DIR_PLUS</includeMacro>
@@ -123,16 +122,16 @@
 									</includeMacros>
 								</configuration>
 							</execution>
-							<!-- winfsp does not support fuse_lowlevel.h but uses fuse2.h -->
+							<!-- winfsp does not support fuse_lowlevel.h. fuse_parse_cmdline is taken from fuse2.h -->
 							<execution>
-								<id>jextract-fuse-parse-cmd-line</id>
+								<id>jextract-fuse-parse-cmdline</id>
 								<goals>
 									<goal>sources</goal>
 								</goals>
 								<configuration>
-									<targetPackage>org.cryptomator.jfuse.win.amd64.extr_fuse2</targetPackage>
+									<targetPackage>org.cryptomator.jfuse.win.amd64.extr.fuse2</targetPackage>
 									<headerFile>${project.parent.basedir}/winfsp/inc/fuse/fuse_common.h</headerFile>
-									<headerClassName>fuse_2_h</headerClassName>
+									<headerClassName>fuse2_h</headerClassName>
 									<includeFunctions>
 										<includeFunction>fuse_parse_cmdline</includeFunction>
 									</includeFunctions>
