+++ jfuse-win-amd64/src/main/java/org/cryptomator/jfuse/win/amd64/extr/fuse2/Constants$root.java
@@ -1,6 +1,6 @@
 // Generated by jextract
 
-package org.cryptomator.jfuse.win.amd64.extr_fuse2;
+package org.cryptomator.jfuse.win.amd64.extr.fuse2;
 
 import java.lang.invoke.MethodHandle;
 import java.lang.invoke.VarHandle;
