+++ jfuse-win-amd64/src/main/java/org/cryptomator/jfuse/win/amd64/extr/fuse2/RuntimeHelper.java
@@ -1,4 +1,4 @@
-package org.cryptomator.jfuse.win.amd64.extr_fuse2;
+package org.cryptomator.jfuse.win.amd64.extr.fuse2;
 // Generated by jextract
 
 import java.lang.foreign.Addressable;
