+++ jfuse-win-amd64/src/main/java/org/cryptomator/jfuse/win/amd64/extr/fuse2/fuse2_h.java
@@ -1,15 +1,15 @@
 // Generated by jextract
 
-package org.cryptomator.jfuse.win.amd64.extr_fuse2;
+package org.cryptomator.jfuse.win.amd64.extr.fuse2;
 
 import java.lang.invoke.MethodHandle;
 import java.lang.invoke.VarHandle;
 import java.nio.ByteOrder;
 import java.lang.foreign.*;
 import static java.lang.foreign.ValueLayout.*;
-public class fuse_2_h  {
+public class fuse2_h  {
 
-    /* package-private */ fuse_2_h() {}
+    /* package-private */ fuse2_h() {}
     public static OfByte C_CHAR = Constants$root.C_CHAR$LAYOUT;
     public static OfShort C_SHORT = Constants$root.C_SHORT$LAYOUT;
     public static OfInt C_INT = Constants$root.C_LONG$LAYOUT;
